requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
pydantic>=2.9.0
gspread
google-auth
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
pandas>=2.1.0
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.2
faiss-cpu
tiktoken

